use gomplate and a yaml config file to configure wireguard
gomplate is a fairly heav dependency, at 10mb, for lightweight images.
But it's extremely flexible compared to expanding templates from a shell
script.  Will use it for now and re-evaluate later.  We can also run
`apk del` on the package after the first boot.

`/etc/init.d/customize_image` is a work in progress.  Still to do:

* After running, delete `config.yaml`.
* Run `lbu commit` to persist expanded templates into overlay tarball.