add simple GitHub Actions setup for continuous integration
Expecting this to fail because qemu will not be setup on the
Ubuntu runner.

Calling `make rpi-basic` and `make rpi-firewall` separately is a
workaround because our Makefile is still rather naive.  Ideally,
`make` would build all the images.  That currently fails because
the first image runs `./bootstrap/armv7/destroy --remove` to
delete the chroot and the second image target does not know to
rebuild the chroot.