k0s: run Argo CD server with `--insecure` flag to support ingress
As [described] in the helm chart, the server needs an `--insecure` flag
to allow TLS termination at the ingress.  Without that setting, Ingress
into port 80 or port 443 of the server results in an infinite HTTP 307
redirect loop.

Similar to https://github.com/argoproj/argo-cd/issues/3169 , but
configuring the URL as `https://argocd.k8s` was insufficient.

It's possible the insecure flag would not be needed if the
`server.certificate.domain` value [is set].

Also updated to the latest release of the Helm chart.  That was
just opportunistic maintenance.

[described]: https://github.com/argoproj/argo-helm/blob/argo-cd-3.26.9/charts/argo-cd/templates/NOTES.txt#L13-L15
[is set]: https://github.com/argoproj/argo-helm/blob/argo-cd-3.26.9/charts/argo-cd/values.yaml#L793-L797