update all images to Alpine 3.15
The [alpine-chroot-install] upgrade is needed for [this commit] where
the Alpine 3.15 signing keys are added to the known key list.

Gotcha to be aware of:  The go ssh library may have trouble connecting
to the new version of openssh-server in Alpine 3.15 unless this line is
added to the `sshd_config`:

    PubkeyAcceptedAlgorithms +ssh-rsa

See https://github.com/golang/go/issues/49952 .  Noticed with [k0sctl]
and the [make build_cluster] target.

[alpine-chroot-install]: https://github.com/alpinelinux/alpine-chroot-install
[k0sctl]: https://github.com/k0sproject/k0sctl
[make build_cluster]: https://github.com/bfritz/homelab-bootstrap/blob/v0.0.3/k0s/Makefile#L52-L53
[this commit]: https://github.com/alpinelinux/alpine-chroot-install/commit/d4ab61fb6e080e3064846e577b18ecab0f0c15ad