rpi-snapcast-client: add Raspberry Pi Zero image to stream audio
Install `snapcast` and configure the `snapcast-client` daemon to
connect to a snapcast server to stream audio from it.  Used as source
for multi-source audio amplifier.

Assumes [HiFiBerry]-compatible DAC is installed to the Raspberry Pi.
Tested with [pHAT DAC].  Edit this section of `usercfg.txt` to use a
different DAC:

    dtparam=audio=on
    dtoverlay=hifiberry-dac

The following configuration variables can be used to customize the
image:

* `HL_WIFI_SSID` - WiFi SSID
* `HL_WIFI_PSK` - WiFi passphrase
* `HL_SNAPCAST_SERVER` - hostname or IP of the snapcast server
* `HL_NTP_SERVER` - NTP server to override the default of `pool.ntp.org`

[HiFiBerry]: https://www.hifiberry.com/dacs
[pHAT DAC]: https://www.adafruit.com/product/3016