rpi-ruuvi: add work-in-progress image for relaying Ruuvi Tag data to InfluxDB
The ruuvitag-listener command installed to `/usr/local/bin` is an
armhf build of lautis' [ruuvitag-listener].  The script to forward
the InfluxDB line protocol data to a server server is not included
in the image.  It's really rough and not 100% reliable.  (That's also
why the tmux and pv packages are included in the image.)

Planning to decommission InfluxDB at some point and may fork
ruuvitag-listener or switch to a different ruuvitag-to-metrics
utility.

[ruuvitag-listener]: https://github.com/lautis/ruuvitag-listener