k0s, k0s-worker, rpi-k0s-controller: use Cilium CNI plugin
Had trouble with the k0s default of kube-router where
prometheus could not connect to TCP 19115 on the
blackbox-exporter pod, despite network policies in place
to allow it.

Began tracing the iptables rule chains, bridges, and other
network magic to find the problem and lost patience before
finding the root cause.

In local VM testing, Cilium seems to work well and has
some built-in diagnostic tools if the connectivity problems
return.