homeassistant: add Home Assistant running in virtualenv
Adds x86 image to run Home Assistant in virtualenv at
`/srv/homeassistant`.

Data is stored to `/data/homeassistant` and there is an init script
to look for a partition with label `ha_data` and mount it at `/data`.
Allows for keeping the Home Assistant database and configuration on
separate partition using more durable media.

Home Assistant 2025.3.4 is in pypi, but using 2025.1.4 because all
newer versions need python 3.13.

Added python, python-dev, git, etc to the Docker build image so the
genapkovl script could setup the virtualenv and install custom
integrations.