homeassistant: bake most pypi modules into overlay
Previously, the Home Assistant container needed HTTP(s) access
to pull packages with pip from pypi.org when the HA service
started. Download and incorporate those packages into the
image, in an apkovl tarball, so the HA host no longer needs
Internet access.

Uses pip constraints and requirements from git repo but disables
several that either won't compile as-is or won't be used and use
a lot of RAM when the apkovl is loaded onto tmpfs.